+++ src/main/php/com-realexpayments-remote-sdk/domain/payment/normaliser/AddressNormaliser.php
@@ -5,6 +5,7 @@ namespace com\realexpayments\remote\sdk\domain\payment\normaliser;
 
 
 use com\realexpayments\remote\sdk\domain\payment\Address;
+use com\realexpayments\remote\sdk\utils\NormaliserHelper;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 class AddressNormaliser implements NormalizerInterface {
@@ -25,7 +26,7 @@ class AddressNormaliser implements NormalizerInterface {
 			'@type'   => $object->getType(),
 			'code'    => $object->getCode(),
 			'country' => $object->getCountry()
-		) );
+		), array( NormaliserHelper::GetClassName(), "filter_data" ) );
 	}
 
 	/**