+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the SDK will be documented in this file.
 ## [1.1]
 - Added mobile payment type (auth-mobile) and relevant fields (mobile, token) to payment request.
 - Added settle, void, rebate, OTB, credit, hold and release payment types.
+- Added Query payment type
  
 
 ## [1.0.1]